Move casper files to tests, upload_captures to tasks/tests.js
