+++ tests/casper-helpers.js
